chore: split credential-provider-ini into multiple files
